refactor(mesh/context): move login context settings to app/context
